projects/oauth2_implementation/requirements.txt
Flask==3.0.0
requests==2.31.0
google-cloud-datastore==2.15.1